Implement Core DB Features and Initial BDD Fixes
This commit implements several core relational database features, fixes numerous bugs, and corrects issues in the BDD test suite.

**Core Feature Implementation:**

- **Indexing:**
  - Enhanced the `Indexer` class to support unique constraints via an `is_unique` flag.
  - Implemented logic to enforce uniqueness on document insertion/update and persist the constraint.
  - Added functionality to bulk-load existing documents into a newly created unique index.

- **Primary & Foreign Keys:**
  - Replaced slow, placeholder O(N) scans for primary and foreign key enforcement with efficient index lookups.
  - Automated the creation of unique indexes for primary keys upon collection creation.
  - Updated `put` and `del` methods to correctly maintain indexes during document modifications.

- **Joins:**
  - Replaced the inefficient nested-loop join algorithm with a more performant index-based join for equi-joins.
  - Added logic to correctly handle `RIGHT` and `FULL` outer joins.

**BDD Test Remediation:**

- Fixed the BDD runner's step matching logic to correctly select the most specific step definition.
- Corrected invalid JSON (improperly quoted keys) within the `database.feature`, `extended_database_tests.feature`, and `more_database_tests.feature` files.
- Fixed a bug in a BDD step to correctly parse collection names from `UPDATE` and `DELETE` TissQL statements.

**Bug Fixes & Compilation:**

- Resolved a critical routing bug in `http_server.cpp` that caused `/_query` requests to fail.
- Made the collection deletion logic idempotent, ensuring it returns the correct HTTP status codes (204 or 404).
- Fixed all provided C++ compilation errors, including declaration mismatches, unused parameter warnings, and signed/unsigned integer comparison warnings.

**Disclaimer:** As per our discussion, these changes could not be verified by running the BDD test suite. The test runner fails to initialize because it cannot find the pre-compiled `tissdb` executable, which I am not permitted to build. The tests will likely fail on your end until the executable is provided to the test environment.

+++ tests/db/http_client.cpp
@@ -39,7 +39,7 @@ int HttpClient::connect_to_server() {
     int sock = -1;
 #ifdef _WIN32
     sock = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
-    if (sock == INVALID_SOCKET) {
+    if (sock == static_cast<int>(INVALID_SOCKET)) {
         throw std::runtime_error("Error creating socket: " + std::to_string(WSAGetLastError()));
     }
 #else
